# src/main/java/com/empresa/serviciofraude/dto/AlertaResponse.java
package com.empresa.serviciofraude.dto;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.math.BigDecimal;
import java.time.Instant;
import java.time.LocalDate;
import java.util.List;
import java.util.UUID;

/**
 * Contrato inmutable Java 21 Record derivado dinámicamente de OpenAPI 3.1
 * Esquema: AlertaResponse
 */
public record AlertaResponse(
    UUID id,

    UUID clienteId,

    String numeroTarjeta,

    @Min(value = 0)
    Integer scoreRiesgo,

    String nivelSeveridad,

    String estado,

    Instant fechaCreacion,

    Instant fechaActualizacion,

    String detalles
) {}